feat: add persistent notification flow with socket delivery
- Add a dedicated socketService with Socket.IO, MongoDB-backed notification storage, auth checks, and service token protection.
- Persist notifications with recipient-based delivery and enrich actor details from authService responses instead of storing names in MongoDB.
- Add notification publishing from managementService for leave creation, approval, and rejection events.
- Route employee leave requests to team leads, team lead leave requests to admins, and approval/rejection results back to the leave owner.
- Add notification API endpoints for listing, marking as read, marking all as read, deleting, and clearing all notifications.
- Add real-time client socket handling for incoming notifications and reuse the same top-right card UI for live and stored notifications.
- Update the sidebar with a notifications entry, badge count, and top-right replay of stored notifications on demand.
- Make top-right dismiss actions mark notifications as read instead of deleting them, while keeping delete actions available through notification management flows.
- Add support for fractional leave durations by changing leave days from integer to float and migrating existing schema accordingly.
- Update Docker Compose and package dependencies for the new socket service and client socket integration.

+++ backend/managementService/models/Leave.js
@@ -19,7 +19,7 @@ module.exports = (sequelize, DataTypes) => {
         type: DataTypes.DATEONLY,
         field: "start_date",
       },
-      days: DataTypes.INTEGER,
+      days: DataTypes.FLOAT,
       reason: DataTypes.STRING,
       details: DataTypes.TEXT,
       status: {
